:sparkles: feat 게임방 조회 rest api 구현

See merge request s07-final/S07P31A301!10

+++ BE/the-knight/src/main/java/com/a301/theknight/domain/game/dto/GameInfoResponse.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class GameInfoResponse {
-    private int gameId;
+    private long gameId;
     private String title;
     private int capacity;
     private int participant;
